refactor(web): [Timer] separate timer functionality into distinct hooks
- Introduced new hooks: `useTimerActions`, `useTimerPlanStatus`, `useTimerStorage`, and `useTimerUi` to modularize timer-related logic.
- Updated existing components to utilize the new hooks, improving code organization and maintainability.
- Removed direct usage of the old `useTimer` hook in favor of specialized hooks for better performance and clarity.
- Enhanced timer state management by ensuring separation of concerns between API interactions, local storage, and UI updates.

This refactor streamlines timer functionality and enhances the overall architecture of the application.
