+++ apps/web/core/hooks/organizations/teams/use-active-team.ts
@@ -4,7 +4,7 @@ import { useCallback } from 'react';
 import { TeamItem } from '@/core/components/teams/team-item';
 import { useTranslations } from 'next-intl';
 import { useCreateOrganizationTeam } from './use-create-organization-team';
-import { useTimer } from '../../activities';
+import { useTimerActions } from '../../timer';
 import { toast } from 'sonner';
 import { useAtomValue } from 'jotai';
 import { activeTeamState } from '@/core/stores';
@@ -12,7 +12,7 @@ import { activeTeamState } from '@/core/stores';
 export const useActiveTeam = () => {
 	const activeTeam = useAtomValue(activeTeamState);
 	const { setActiveTeam } = useCreateOrganizationTeam();
-	const { timerStatus, stopTimer } = useTimer();
+	const { timerStatus, stopTimer } = useTimerActions();
 	const t = useTranslations();
 	const onChangeActiveTeam = useCallback(
 		(item: TeamItem) => {
